+++ html4docx/__init__.py
@@ -1,1 +1,4 @@
+import logging
 from .h4d import HtmlToDocx
+
+logging.getLogger(__name__).addHandler(logging.NullHandler())